# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Community Growth Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ErrorResponse {
  error: string;
}

export interface WhatsAppStatus {
  connected: boolean;
  status: string;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  displayName?: string | null;
}

export interface WhatsAppQR {
  /** @nullable */
  qr?: string | null;
  status: string;
}

export interface Contact {
  id: number;
  name: string;
  phone: string;
  tags?: string[] | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  tags?: string[] | null;
  /** @nullable */
  notes?: string | null;
}

export interface ContactUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  phone?: string;
  tags?: string[] | null;
  /** @nullable */
  notes?: string | null;
}

export interface BulkContactInput {
  text: string;
}

export interface BulkImportResult {
  imported: number;
  skipped: number;
  errors: string[];
}

export interface BulkDeleteInput {
  ids: number[];
}

export interface Group {
  id: number;
  groupId: string;
  name: string;
  /** @nullable */
  memberCount?: number | null;
  createdAt: string;
}

export interface GroupInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  groupId?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  memberCount?: number | null;
}

export interface AddContactsInput {
  contactIds: number[];
}

export interface Campaign {
  id: number;
  name: string;
  messageTemplate: string;
  status: string;
  /** @nullable */
  startDate?: string | null;
  createdAt: string;
  sentCount: number;
  failedCount: number;
  totalCount: number;
}

export interface CampaignInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  messageTemplate: string;
  contactIds: number[];
}

export interface CampaignUpdate {
  name?: string;
  messageTemplate?: string;
  status?: string;
}

export interface MessageLog {
  id: number;
  /** @nullable */
  contactId?: number | null;
  /** @nullable */
  campaignId?: number | null;
  /** @nullable */
  contactName?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  /** @nullable */
  recipientPhone?: string | null;
  /** @nullable */
  campaignName?: string | null;
  status: string;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  sentAt?: string | null;
  createdAt: string;
}

export interface GroupLog {
  id: number;
  /** @nullable */
  contactId?: number | null;
  /** @nullable */
  groupId?: number | null;
  /** @nullable */
  contactName?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  /** @nullable */
  groupName?: string | null;
  /** @nullable */
  action?: string | null;
  /** @nullable */
  details?: string | null;
  status: string;
  createdAt: string;
}

export interface DashboardStats {
  totalContacts: number;
  totalGroups: number;
  messagesSentToday: number;
  activeCampaigns: number;
  successRate: number;
  totalMessagesSent: number;
  totalMessagesFailed: number;
}

export interface ActivityItem {
  id: string;
  type: string;
  message: string;
  timestamp: string;
  /** @nullable */
  status?: string | null;
}

export interface ComposeMessageInput {
  topic: string;
  /** @nullable */
  tone?: string | null;
  /** @nullable */
  purpose?: string | null;
}

export interface ComposeMessageResult {
  message: string;
}

export interface MessageTemplate {
  id: number;
  name: string;
  content: string;
  /** @nullable */
  tone?: string | null;
  /** @nullable */
  purpose?: string | null;
  createdAt: string;
}

export interface MessageTemplateInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  content: string;
  /** @nullable */
  tone?: string | null;
  /** @nullable */
  purpose?: string | null;
}

export type ListContactsParams = {
search?: string;
};

export type ListMessageLogsParams = {
status?: string;
};

export type ListGroupLogsParams = {
status?: string;
};