# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhatsApp Community Growth Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all campaigns
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /campaigns/{id}/start:
    post:
      operationId: startCampaign
      tags: [campaigns]
      summary: Start a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /campaigns/{id}/pause:
    post:
      operationId: pauseCampaign
      tags: [campaigns]
      summary: Pause a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /campaigns/{id}/cancel:
    post:
      operationId: cancelCampaign
      tags: [campaigns]
      summary: Cancel a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /campaigns/{id}/logs:
    get:
      operationId: getCampaignLogs
      tags: [campaigns]
      summary: Get message logs for a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessageLog"

  # Message logs
...
  /ai/compose:
    post:
      operationId: composeMessage
      tags: [ai]
      summary: Use AI to personalize/rewrite a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComposeMessageInput"
      responses:
        "200":
          description: AI-composed message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComposeMessageResult"
  /ai/templates:
    get:
      operationId: listTemplates
      tags: [ai]
      summary: List message templates
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessageTemplate"
    post:
      operationId: createTemplate
      tags: [ai]
      summary: Create a message template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageTemplateInput"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageTemplate"
  /ai/templates/{id}:
    delete:
      operationId: deleteTemplate
      tags: [ai]
      summary: Delete a message template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        messageTemplate:
          type: string
        status:
          type: string
        startDate:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        sentCount:
          type: integer
        failedCount:
          type: integer
        totalCount:
          type: integer
      required:
        - id
        - name
        - messageTemplate
        - status
        - createdAt
        - sentCount
        - failedCount
        - totalCount
    CampaignInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        messageTemplate:
          type: string
          minLength: 1
        contactIds:
          type: array
          items:
            type: integer
      required:
        - name
        - messageTemplate
        - contactIds
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
        messageTemplate:
          type: string
        status:
          type: string
    MessageLog:
      type: object
      properties:
        id:
          type: integer
        contactId:
          type: ["integer", "null"]
        campaignId:
          type: ["integer", "null"]
        contactName:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        recipientPhone:
          type: ["string", "null"]
        campaignName:
          type: ["string", "null"]
        status:
          type: string
        message:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        sentAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - status
        - createdAt
...
    ComposeMessageInput:
      type: object
      properties:
        topic:
          type: string
        tone:
          type: ["string", "null"]
        purpose:
          type: ["string", "null"]
      required:
        - topic
    ComposeMessageResult:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    MessageTemplate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        content:
          type: string
        tone:
          type: ["string", "null"]
        purpose:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - content
        - createdAt
    MessageTemplateInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        tone:
          type: ["string", "null"]
        purpose:
          type: ["string", "null"]
      required:
        - name
        - content