# lib/api-zod/src/generated/types/messageTemplate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Community Growth Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface MessageTemplate {
  id: number;
  name: string;
  content: string;
  /** @nullable */
  tone?: string | null;
  /** @nullable */
  purpose?: string | null;
  createdAt: Date;
}